rustc_codegen_ssa: move debuginfo-related things to a new mir::debuginfo module.
